feat: zustand setup for http status list
