+++ examples/kitchen-sink/statuscodes/web/react-shadcn-tailwind-zustand/package.json
@@ -15,6 +15,7 @@
     "@radix-ui/react-navigation-menu": "^1.1.4",
     "@radix-ui/react-select": "^2.0.0",
     "@radix-ui/react-slot": "^1.0.2",
+    "@redux-devtools/extension": "^3.3.0",
     "accordion": "^3.0.2",
     "add": "^2.0.6",
     "class-variance-authority": "^0.7.0",
@@ -26,7 +27,7 @@
     "shadcn-ui": "^0.8.0",
     "tailwind-merge": "^2.2.1",
     "tailwindcss-animate": "^1.0.7",
-    "zustand": "^4.5.0"
+    "zustand": "^4.5.2"
   },
   "devDependencies": {
     "@types/node": "^20.11.17",
