+++ examples/kitchen-sink/statuscodes/web/react-shadcn-tailwind-zustand/src/constants/types.ts
@@ -1,0 +1,4 @@
+
+export interface MappingType {
+  [key: string]: string;
+}
