+++ examples/kitchen-sink/statuscodes/web/react-shadcn-tailwind-zustand/src/store/statusCodes.store.ts
@@ -1,0 +1,12 @@
+import { StatusListType } from '@/pages/codesList/components/types';
+import { create } from 'zustand'
+
+interface StatusCodes {
+    HTTPStatusCodesList: StatusListType;
+    setHTTPStatusCodesList: (HTTPStatusCodesList: StatusListType) => void;
+}
+
+export const useStatusCodeStore = create<StatusCodes>()((set) => ({
+  HTTPStatusCodesList:{},
+  setHTTPStatusCodesList: (HTTPStatusCodesList) => set({ HTTPStatusCodesList }),
+}))
